node backend and Admin window: major refactor
# Node:

handshake()
- creates and returns a status object
- status object includes error codes and long messages for all errors
- updates many new attributes save in main kiosk settings file

  - cfa_registered: false,
  - cfa_credential_valid: false,
  - cfa_ip_address_valid: false,
  - ip_address: false,
  - online: false,
  - working: false,
  - autostart_visitor: true

Calls from renderer process using ipc.Renderer.invoke()
now ignore any return values from ipcMain.handle().
Instead objects from node process are sent to the Admin
renderer using adminWindow.webContents.send() and received by:

- window.api.kioskStateUpdate()
- window.api.kioskLogStateUpdate()
- window.api.kioskStatusUpdate()

These functions are exposed in the window object in preload.js.

Admin window:

# Autostart Visitor application

Autostart the Visitor application 60s after the Admin app
is started, This autostart will be either disabled or stopped
any of the following condition or events occurr.

1. Any user interaction with the page.
2. If the CfA communication status is (or becomes) invalid.

*** TODO: actually start the Visitor application.

# CfA communication status indicators

A set of CfA communication status indicators showing valid or
invalid states are displayed. The indicators consist of:

- online
- registered
- credential
- ipaddress

In addition a 'working' indicator is displayed. For this to
be valid all of the other conditions need to be valid.

These status attributes are saved in the main application settings.

If communication with CfA is not possible for any of the conditions
a detailed message is displayed below the indicators.

*** TODO: expose and use the status attributes in the main
Visitor application

# Handshake

A disclosure triangle is available to expand or contract
the logs showing the detailed results.