image devtool panel: improve entering values < 1
Example: 

When trying to change a min value set as 0.1 to 0.05 it is natural
first delete the '1' char and then type '05'. As soon as the '1'
character is deleted the new value of the min setting is set to zero,
the page is re-rendered and the displayed value in the min field is
reset to '0'.

After this change:

When characters are entered into the field the value is only
persisted and page rendered if the new value is different
from the persisted value.