admin: wait 250ms after online-status changes
An experiment to possibly fix this bug on the windows kisok systems:

When Admin window open and power to computer is cut the settings file
cfa.json ends up empty.

Guess is that when power is abruptly removed the network interface goes
down before the cpu shuts down. The change in on-line status is routed
to the nodejs backend where the handler immediately starts to update
cfa.json with the new status for online. cpu shuts down in the middle
of this process while file is open and starting to be written to.